S7V3 everything working

+++ section-7/video-3/api-gateway-1/src/handlers/getAllUserVideos.go
@@ -11,7 +11,7 @@ func (h *Handler) GetAllUserVideos(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	userID := ctx.Value(UserID).(uint32)
 
-	allUserVideos, err := h.GetAllUserVideosUC.GetAllVideosFromUser(uint32(userID))
+	allUserVideos, err := h.GetAllUserVideosUC.GetAllVideosFromUser(userID)
 
 	if err != nil {
 		w.WriteHeader(http.StatusInternalServerError)
